+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Data to connect the network in metamask
 - Full balance tooltip
 - Bids list logic parachains auction tab
+- Added webpack chunks 
 
 #### Changed
 - Version api packages